resamp: fix memory corruption
Passing the output buffer size to swr_convert in bytes when it
expects samples made it think more output space was available
than had been allocated.

When it tried to use the extra space (ie. when resampling to a
higher sample rate), it ended up clobbering all kinds of memory.